Add tests checking if the synchronous call mechanism is working properly
